No long long in Qt 3.1.
git-svn-id: http://svn.mythtv.org/svn/trunk@5371 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/commercial_skip.cpp
@@ -1320,7 +1320,7 @@ void CommDetect::BuildAllMethodsCommList(void)
                             "block %1 and going to end of program. length "
                             "is %2 frames")
                             .arg(curBlock)
-                            .arg(framesProcessed - breakStart - 1));
+                            .arg((long)(framesProcessed - breakStart - 1)));
 
         commBreakMap[breakStart] = MARK_COMM_START;
         commBreakMap[framesProcessed - 1] = MARK_COMM_END;
